fixes by review
