+++ KODE Weather App/Models/Response.swift
@@ -1,0 +1,10 @@
+//
+//  Response.swift
+
+import Foundation
+
+struct Response: Decodable {
+    var weather: [WeatherBlock]
+    var main: MainBlock
+    var wind: WindBlock
+}
