+++ KODE Weather App/Models/WeatherBlock.swift
@@ -1,0 +1,11 @@
+//
+//  WeatherBlock.swift
+
+import Foundation
+
+struct WeatherBlock: Decodable {
+    var id: Int
+    var main: String
+    var description: String
+    var icon: String
+}
