+++ KODE Weather App/Models/WindBlock.swift
@@ -1,0 +1,9 @@
+//
+//  WindBlock.swift
+
+import Foundation
+
+struct WindBlock: Decodable {
+    var speed: Double
+    var deg: Double
+}
