+++ KODE Weather App/Protocols/Declarations/WeatherNetworkService.swift
@@ -1,0 +1,8 @@
+//
+//  WeatherNetworkService.swift
+
+import PromiseKit
+
+protocol WeatherNetworkService {
+    func getWeather(city: String) -> Promise<Response>
+}
