--- KODE Weather App/Values/Constants.swift
@@ -1,14 +1,0 @@
-//
-//  Constants.swift
-//  KODE Weather App
-//
-//  Created by Евгений Урбановский on 17.07.2020.
-//  Copyright © 2020 Evgeniy. All rights reserved.
-//
-
-import Foundation
-
-enum Constants {
-    static let url = "http://api.openweathermap.org/"
-    static let APIKey = "9fdb3ef0cca5b3b8afe78aff22fd7eef"
-}
