+++ KODE Weather App/Values/URLFactory.swift
@@ -1,0 +1,8 @@
+//
+//  Constants.swift
+
+import Foundation
+
+struct URLFactory {
+    static let url = "http://api.openweathermap.org/"
+}
